marvinai/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-hosts==4.0
django-js-asset==1.2.2
django-maskurl==0.1
django-mptt==0.11.0
dnsimple==1.2.0
gunicorn==20.0.4
idna==2.9
numpy==1.18.4
pandas==1.0.4
pkg-resources==0.0.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
